# djobea_ai_clean_swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Djobea AI",
    "description": "Agent conversationnel WhatsApp pour services \u00e0 domicile au Cameroun",
    "version": "1.0.0"
  },
  "paths": {
    "/webhook/whatsapp": {
      "post": {
        "tags": [
          "webhook"
        ],
        "summary": "Whatsapp Webhook",
        "description": "Handle incoming WhatsApp messages",
        "operationId": "whatsapp_webhook_webhook_whatsapp_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_whatsapp_webhook_webhook_whatsapp_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/system": {
      "get": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Get System Settings",
        "description": "GET /api/settings/system - Get system settings\nRetrieve system-wide configuration settings",
        "operationId": "get_system_settings_api_settings_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Update System Settings",
        "description": "PUT /api/settings/system - Update system settings\nUpdate system-wide configuration settings",
        "operationId": "update_system_settings_api_settings_system_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Settings Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/settings/notifications": {
      "get": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Get Notification Settings",
        "description": "GET /api/settings/notifications - Get notification settings\nRetrieve notification configuration and templates",
        "operationId": "get_notification_settings_api_settings_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Update Notification Settings",
        "description": "PUT /api/settings/notifications - Update notification settings\nUpdate notification configuration and templates",
        "operationId": "update_notification_settings_api_settings_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Notification Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/settings/pricing": {
      "get": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Get Pricing Settings",
        "description": "GET /api/settings/pricing - Get pricing settings\nRetrieve pricing configuration and service rates",
        "operationId": "get_pricing_settings_api_settings_pricing_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Update Pricing Settings",
        "description": "PUT /api/settings/pricing - Update pricing settings\nUpdate pricing configuration and service rates",
        "operationId": "update_pricing_settings_api_settings_pricing_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Pricing Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/settings/integrations": {
      "get": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Get Integration Settings",
        "description": "GET /api/settings/integrations - Get integration settings\nRetrieve external service integrations configuration",
        "operationId": "get_integration_settings_api_settings_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings",
          "Settings"
        ],
        "summary": "Update Integration Settings",
        "description": "PUT /api/settings/integrations - Update integration settings\nUpdate external service integrations configuration",
        "operationId": "update_integration_settings_api_settings_integrations_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Integration Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_whatsapp_webhook_webhook_whatsapp_post": {
        "properties": {
          "AccountSid": {
            "type": "string",
            "title": "Accountsid"
          },
          "MessageSid": {
            "type": "string",
            "title": "Messagesid"
          },
          "From": {
            "type": "string",
            "title": "From"
          },
          "To": {
            "type": "string",
            "title": "To"
          },
          "Body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "NumMedia": {
            "type": "integer",
            "title": "Nummedia",
            "default": 0
          },
          "MediaUrl0": {
            "type": "string",
            "title": "Mediaurl0"
          },
          "MediaContentType0": {
            "type": "string",
            "title": "Mediacontenttype0"
          }
        },
        "type": "object",
        "required": [
          "AccountSid",
          "MessageSid",
          "From",
          "To"
        ],
        "title": "Body_whatsapp_webhook_webhook_whatsapp_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}